Moved test data builders
(and fixed IDEA inspections)
